API Reference

StringAnalysis.StringAnalysis

StringAnalysis.LSAModel

StringAnalysis.TextHashFunction

Base.size

Base.summary

Base.summary

StringAnalysis.cosine

StringAnalysis.embed_document

StringAnalysis.embed_word

StringAnalysis.get_vector

StringAnalysis.in_vocabulary

StringAnalysis.index

StringAnalysis.lda

StringAnalysis.load

StringAnalysis.lsa

StringAnalysis.remove_patterns

StringAnalysis.remove_patterns!

StringAnalysis.save

StringAnalysis.sentence_tokenize

StringAnalysis.similarity

StringAnalysis.tokenize

StringAnalysis.tokenize_fast

StringAnalysis.tokenize_slow

StringAnalysis.vocabulary

StringAnalysis.StringAnalysis — Module.

A Julia library for working with text hard-forked from TextAnalysis.jl.

source

StringAnalysis.LSAModel — Type.

LSAModel{S<:AbstractString, T<:AbstractFloat, A<:AbstractMatrix{T}, H<:Integer}

LSA (latent semantic analysis) model. It constructs from a document term matrix (dtm) a model that can be used to embed documents in a latent semantic space pertaining to the data. The model requires that the document term matrix be a DocumentTermMatrix{T<:AbstractFloat} because the matrices resulted from the SVD operation will be forced to contain elements of type T.

Fields

vocab::Vector{S} a vector with all the words in the corpus

vocab_hash::Dict{S,H} a word to index in word embeddings matrix mapping

U::A the document embeddings matrix

Σinv::A inverse of the singular value matrix

Vᵀ::A transpose of the word embedding matrix

stats::Symbol the statistical measure to use for word importances in documents available values are: :tf (term frequency) :tfidf (default, term frequency - inverse document frequency) :bm25 (Okapi BM25)

idf::Vector{T} inverse document frequencies for the words in the vocabulary

nwords::T averge number of words in a document

κ::Int the κ parameter of the BM25 statistic

β::Float64 the β parameter of the BM25 statistic

tol::T minimum size of the vector components (default T(1e-15))

U, Σinv and Vᵀ:

If X is a m×n document-term-matrix with m documents and n words so that X[i,j] represents a statistical indicator of the importance of term j in document i then:

U, Σ, V = svd(X)

Σinv = inv(Σ)

Vᵀ = V'

The version of U actually stored in the model has its columns normalized to their norm.

Examples

julia> using StringAnalysis

       doc1 = StringDocument("This is a text about an apple. There are many texts about apples.")
       doc2 = StringDocument("Pears and apples are good but not exotic. An apple a day keeps the doctor away.")
       doc3 = StringDocument("Fruits are good for you.")
       doc4 = StringDocument("This phrase has nothing to do with the others...")
       doc5 = StringDocument("Simple text, little info inside")

       crps = Corpus(AbstractDocument[doc1, doc2, doc3, doc4, doc5])
       prepare!(crps, strip_punctuation)
       update_lexicon!(crps)
       dtm = DocumentTermMatrix{Float32}(crps, sort(collect(keys(crps.lexicon))))

       ### Build LSA Model ###
       lsa_model = LSAModel(dtm, k=3, stats=:tf)

       query = StringDocument("Apples and an exotic fruit.")
       idxs, corrs = cosine(lsa_model, query)

       println("Query: "$(query.text)"")
       for (idx, corr) in zip(idxs, corrs)
           println("$corr -> "$(crps[idx].text)"")
       end
Query: "Apples and an exotic fruit."
0.9746108 -> "Pears and apples are good but not exotic  An apple a day keeps the doctor away "
0.870703 -> "This is a text about an apple  There are many texts about apples "
0.7122063 -> "Fruits are good for you "
0.22725986 -> "This phrase has nothing to do with the others "
0.076901935 -> "Simple text  little info inside "

References:

The LSA wiki page

Deerwester et al. 1990

source

StringAnalysis.TextHashFunction — Type.

TextHashFunction(hash_function::Function, cardinality::Int)

The basic structure for performing text hashing: uses the hash_function to generate feature vectors of length cardinality.

Details

The hash trick is the use a hash function instead of a lexicon to determine the columns of a DocumentTermMatrix-like encoding of the data. To produce a DTM for a Corpus for which we do not have an existing lexicon, we need someway to map the terms from each document into column indices. We use the now standard "Hash Trick" in which we hash strings and then reduce the resulting integers modulo N, which defines the numbers of columns we want our DTM to have. This amounts to doing a non-linear dimensionality reduction with low probability that similar terms hash to the same dimension.

To make things easier, we wrap Julia's hash functions in a new type, TextHashFunction, which maintains information about the desired cardinality of the hashes.

References:

The "Hash Trick" wiki page

Moody, John 1989

Examples

julia> doc = StringDocument("this is a text")
       thf = TextHashFunction(hash, 13)
       hash_dtv(doc, thf, Float16)
13-element Array{Float16,1}:
 1.0
 1.0
 0.0
 0.0
 0.0
 0.0
 0.0
 2.0
 0.0
 0.0
 0.0
 0.0
 0.0

source

StringAnalysis.cosine — Function.

cosine(lm, doc, n=10)

Return the position of n (by default n = 10) neighbors of document doc and their cosine similarities.

source

StringAnalysis.embed_document — Method.

embed_document(lm, doc)

Return the vector representation of a document doc using the LSA model lm.

source

StringAnalysis.embed_word — Method.

embed_word(lm, word)

Return the vector representation of word using the LSA model lm.

source

StringAnalysis.get_vector — Method.

get_vector(lm, word)

Returns the vector representation of word from the LSA model lm.

source

StringAnalysis.index — Method.

index(lm, word)

Return the index of word from the LSA model lm.

source

StringAnalysis.lda — Method.

ϕ, θ = lda(dtm::DocumentTermMatrix, ntopics::Int, iterations::Int, α::Float64, β::Float64)

Perform Latent Dirichlet allocation.

Arguments

α Dirichlet dist. hyperparameter for topic distribution per document. α<1 yields a sparse topic mixture for each document. α>1 yields a more uniform topic mixture for each document.

β Dirichlet dist. hyperparameter for word distribution per topic. β<1 yields a sparse word mixture for each topic. β>1 yields a more uniform word mixture for each topic.

Return values

ϕ: ntopics × nwords Sparse matrix of probabilities s.t. sum(ϕ, 1) == 1

θ: ntopics × ndocs Dense matrix of probabilities s.t. sum(θ, 1) == 1

source

StringAnalysis.load — Method.

load(filename, type; [sparse=true])

Loads an LSA model from filename into an LSA model object. The embeddings matrix element type is specified by type (default Float32) while the keyword argument sparse specifies whether the matrix should be sparse or not.

source

StringAnalysis.lsa — Method.

lsa(X [;k=3, stats=:tfidf, κ=2, β=0.75, tol=1e-15])

Constructs an LSA model. The input X can be a Corpus or a DocumentTermMatrix. Use ?LSAModel for more details. Vector components smaller than tol will be zeroed out.

source

StringAnalysis.save — Method.

save(lm, filename)

Saves an LSA model lm to disc in file filename.

source

StringAnalysis.sentence_tokenize — Method.

sentence_tokenize([lang,] s)

Splits string s into sentences using WordTokenizers.split_sentences function to perform the tokenization. If a language lang is provided, it ignores it ;)

source

StringAnalysis.similarity — Method.

similarity(lm, doc1, doc2)

Return the cosine similarity value between two documents doc1 and doc2.

source

StringAnalysis.tokenize — Method.

tokenize(s [;method])

Tokenizes based on either the tokenize_slow or tokenize_fast functions.

source

StringAnalysis.tokenize_fast — Method.

tokenize_fast(doc [;splitter])

Function that quickly tokenizes doc based on the splitting pattern specified by splitter::RegEx. Supported types for doc are: AbstractString, Vector{AbstractString}, StringDocument and NGramDocument.

source

StringAnalysis.tokenize_slow — Method.

tokenize_slow([lang,] s)

Splits string s into tokens on whitespace using WordTokenizers.tokenize function to perform the tokenization. If a language lang is provided, it ignores it ;)

source

Base.size — Method.

size(lm)

Return a tuple containing the number of terms, the number of documents and the vector representation dimensionality of the LSA model lm.

source

Base.summary — Method.

summary(doc)

Shows information about the document doc.

source

Base.summary — Method.

summary(crps)

Shows information about the corpus crps.

source

StringAnalysis.in_vocabulary — Method.

in_vocabulary(lm, word)

Return true if word is part of the vocabulary of the LSA model lm and false otherwise.

source

StringAnalysis.remove_patterns! — Method.

remove_patterns!(d, rex)

Removes from the document or corpus d the text matching the pattern described by the regular expression rex.

source

StringAnalysis.remove_patterns — Method.

remove_patterns(s, rex)

Removes from the string s the text matching the pattern described by the regular expression rex.

source

StringAnalysis.vocabulary — Method.

vocabulary(lm)

Return the vocabulary as a vector of words of the LSA model lm.

source